+++ vocoder/hifigan/models.py
@@ -3,7 +3,7 @@ import torch.nn.functional as F
 import torch.nn as nn
 from torch.nn import Conv1d, ConvTranspose1d, AvgPool1d, Conv2d
 from torch.nn.utils import weight_norm, remove_weight_norm, spectral_norm
-from hifigan.utils import init_weights, get_padding
+from vocoder.hifigan.utils import init_weights, get_padding
 
 LRELU_SLOPE = 0.1
