+++ vocoder/wavernn/gen_wavernn.py
@@ -1,5 +1,5 @@
-from vocoder.models.fatchord_version import  WaveRNN
-from vocoder.audio import *
+from vocoder.wavernn.models.fatchord_version import  WaveRNN
+from vocoder.wavernn.audio import *
 
 
 def gen_testset(model: WaveRNN, test_set, samples, batched, target, overlap, save_path):
