+++ vocoder/wavernn/models/fatchord_version.py
@@ -3,7 +3,7 @@ import torch.nn as nn
 import torch.nn.functional as F
 from vocoder.distribution import sample_from_discretized_mix_logistic
 from vocoder.display import *
-from vocoder.audio import *
+from vocoder.wavernn.audio import *
 
 
 class ResBlock(nn.Module):
